+++ changelog.md
@@ -8,6 +8,7 @@
 - Add new variables for version string template: `{git_repo_path}`, `{git_branch}`, `{git_long_hash}`, `{git_short_hash}` and `{git_date}`.
 - Add a new `--no-config` option on  `@extra_command` and `@extra_group` to disable configuration files. Closes {issue}`750`.
 - Add `--table-format` option by default on `@extra_command` and `@extra_group`.
+- Let `--table-format` and `--no-color` affect the rendering of `--show-params` table.
 - Swap `Class` and `Spec.` columns in `--show-params` output.
 - Remove the `-C` short option associated with `--config`.
 - Remove the `-t` short option associated with `--table-format`.